p4i integration Integrate 259331 from 4.3 to main: Fixes:    Streaming incorrect qreal size for opacity on greenphone. RevBy:    Michael Brasser Details:  Make sure we stream opacity as a double on all platforms.
[git-p4: depot-paths = "//depot/qt/main/": change = 259332]

+++ src/gui/image/qpaintengine_pic.cpp
@@ -169,7 +169,7 @@ void QPicturePaintEngine::updateOpacity(qreal opacity)
 #endif
     int pos;
     SERIALIZE_CMD(QPicturePrivate::PdcSetOpacity);
-    d->s << opacity;
+    d->s << double(opacity);
     writeCmdLength(pos, QRectF(), false);
 }
